# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Fincava — Colombian Agricultural B2B Marketplace API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /buyer/orders:
    get:
      operationId: listBuyerOrders
      tags: [orders]
      summary: List buyer's orders
      responses:
        "200":
          description: Orders list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Order"

    post:
      operationId: createOrder
      tags: [orders]
      summary: Create an order
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateOrderBody"
      responses:
        "201":
          description: Order created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
  /buyer/orders/{id}:
    get:
      operationId: getBuyerOrder
      tags: [orders]
      summary: Get order detail
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Order detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrderDetail"
...
  /supplier/orders/{id}/status:
    patch:
      operationId: updateOrderStatus
      tags: [orders]
      summary: Update order status (supplier only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateOrderStatusBody"
      responses:
        "200":
          description: Updated order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"

  # USER PROFILE
...
components:
  schemas:
...
    Order:
      type: object
      properties:
        id:
          type: integer
        buyerId:
          type: integer
        buyerName:
          type: string
        status:
          type: string
          enum: [INQUIRY, SAMPLE_REQUESTED, QUOTED, CONFIRMED, IN_PRODUCTION, SHIPPED, DELIVERED, COMPLETED, CANCELLED]
        totalUSD:
          type: number
        incoterm:
          type: string
        destinationPort:
          type: ["string", "null"]
        shippingMethod:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        itemCount:
          type: integer
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - buyerId
        - buyerName
        - status
        - totalUSD
        - incoterm
        - itemCount
        - createdAt
        - updatedAt
    OrderDetail:
      allOf:
        - $ref: "#/components/schemas/Order"
        - type: object
          properties:
            items:
              type: array
              items:
                $ref: "#/components/schemas/OrderItem"
            messages:
              type: array
              items:
                $ref: "#/components/schemas/Message"
          required:
            - items
            - messages
    OrderItem:
      type: object
      properties:
        id:
          type: integer
        productId:
          type: integer
        productName:
          type: string
        productImage:
          type: ["string", "null"]
        quantityKg:
          type: number
        pricePerKg:
          type: number
        totalUSD:
          type: number
      required:
        - id
        - productId
        - productName
        - quantityKg
        - pricePerKg
        - totalUSD
    CreateOrderBody:
      type: object
      properties:
        incoterm:
          type: string
        destinationPort:
          type: ["string", "null"]
        shippingMethod:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        items:
          type: array
          items:
            $ref: "#/components/schemas/CreateOrderItemBody"
      required:
        - items
    CreateOrderItemBody:
      type: object
      properties:
        productId:
          type: integer
        quantityKg:
          type: number
      required:
        - productId
        - quantityKg
    UpdateOrderStatusBody:
      type: object
      properties:
        status:
          type: string
          enum: [INQUIRY, SAMPLE_REQUESTED, QUOTED, CONFIRMED, IN_PRODUCTION, SHIPPED, DELIVERED, COMPLETED, CANCELLED]
      required:
        - status
...
    Message:
      type: object
      properties:
        id:
          type: integer
        senderId:
          type: integer
        senderName:
          type: string
        receiverId:
          type: integer
        content:
          type: string
        read:
          type: boolean
        createdAt:
          type: string
      required:
        - id
        - senderId
        - senderName
        - receiverId
        - content
        - read
        - createdAt
...