# lib/api-zod/src/generated/types/createOrderBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Fincava — Colombian Agricultural B2B Marketplace API
 * OpenAPI spec version: 0.1.0
 */
import type { CreateOrderItemBody } from "./createOrderItemBody";

export interface CreateOrderBody {
  incoterm?: string;
  /** @nullable */
  destinationPort?: string | null;
  /** @nullable */
  shippingMethod?: string | null;
  /** @nullable */
  notes?: string | null;
  items: CreateOrderItemBody[];
}