# lib/api-zod/src/generated/types/order.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Fincava — Colombian Agricultural B2B Marketplace API
 * OpenAPI spec version: 0.1.0
 */
import type { OrderStatus } from "./orderStatus";

export interface Order {
  id: number;
  buyerId: number;
  buyerName: string;
  status: OrderStatus;
  totalUSD: number;
  incoterm: string;
  /** @nullable */
  destinationPort?: string | null;
  /** @nullable */
  shippingMethod?: string | null;
  /** @nullable */
  notes?: string | null;
  itemCount: number;
  createdAt: string;
  updatedAt: string;
}